refactor continuing
