+++ pytc_gui/main_window.py
@@ -5,160 +5,17 @@ pytc GUI using PyQt5.
 __author__ = "Hiranmyai Duvvuri"
 __date__ = "2017-01-06"
 
+from . import dialog
+from .main_widgets import GUIMaster
+
 from pytc.global_fit import GlobalFit
 
 from PyQt5.QtGui import *
 from PyQt5.QtCore import *
 from PyQt5.QtWidgets import *
 
-from . import dialog
-from .boxes import PlotBox, ExperimentBox
-
-from .qlogging_handler import OutputStream
-
 from matplotlib.backends.backend_pdf import PdfPages
-
-import sys, pkg_resources, pickle, inspect, copy
-
-class GUIMaster(QWidget):
-    """
-    Main class that holds all of the fitter sub-widgets.
-    """
-
-    def __init__(self, parent):
-
-        super().__init__()
-
-        self._parent = parent
-
-        self._fitter = parent._fitter
-        self._fitter_list = parent._fitter_list
-
-        # Create a dictionary of GlobalFitOptions
-        fit_args = inspect.getargspec(GlobalFit().fit)
-        self._global_fit_options = {arg: param for arg, param
-                                    in zip(fit_args.args[1:],fit_args.defaults)}
-
-        # Lay out the widget.
-        self.layout()
-
-    def layout(self):
-        """
-        Create the widget layout.
-        """
-
-        # ------------ Plot widget ----------------------- 
-        self._plot_box = PlotBox(self)
-
-        # ------------ Experiments widget ----------------
-        self._exp_box = ExperimentBox(self)
-    
-        # ------------ "Do fit" button -------------------
-        do_fit_button = QPushButton("Do fit", self)
-        do_fit_button.clicked.connect(self.do_fit_callback)
-
-        # -------------- message box ---------------------
-        self._message_box = QTextEdit()
-        self._message_box.setReadOnly(True)
-
-        scroll = QScrollArea(self)
-        scroll.setWidget(self._message_box)
-        scroll.setWidgetResizable(True)
-
-        # redirect stdout to the message box
-        self._temp = sys.stdout
-        sys.stdout = OutputStream()
-        sys.stdout.text_printed.connect(self.read_stdout_callback)
-
-        # Split up the main window in a useful way
-
-        # Split window vertically
-        v_splitter = QSplitter(Qt.Vertical)
-        v_splitter.addWidget(self._plot_box)
-        v_splitter.addWidget(scroll)
-        v_splitter.setSizes([300, 50])
-
-        # now split horizontally
-        h_splitter = QSplitter(Qt.Horizontal)
-        h_splitter.addWidget(v_splitter)
-        h_splitter.addWidget(self._exp_box)
-        h_splitter.setSizes([200, 200])
-
-        # Now add the split up window.
-        main_layout = QVBoxLayout(self)
-        main_layout.addWidget(h_splitter)
-        main_layout.addWidget(do_fit_button)
-
-        # MJH ??? --> what is signaling architecture?
-        self._parent.fit_signal.connect(self.fit_signal_update)
-
-    def clear(self):
-        """
-        Clear the plot and experiment frames.
-        """
-
-        self._plot_box.clear()
-        self._exp_box.clear()
-
-    def update_fit_options(self, options_dict):
-        """
-        """
-        self._global_fit_options = options_dict
-
-    def do_fit_callback(self):
-        """
-        """
-        self._exp_box.perform_fit(self._global_fit_options)
-        self._plot_box.update()
-
-    @pyqtSlot(GlobalFit)
-    def fit_signal_update(self, obj):
-        """
-        """
-        self._plot_box._fitter = obj
-        self._exp_box._fitter = obj
-
-    @pyqtSlot(str)
-    def read_stdout_callback(self, text):
-        """
-        Wirte standard out to the main message box.
-        """
-        self._message_box.insertPlainText(text)
-
-    @property
-    def parent(self):
-        """
-        Parent window.
-        """
-        return self._parent
-
-    @property
-    def plot_box(self):
-        """
-        Main plot box.
-        """ 
-        return self._plot_box
-
-    @property
-    def exp_box(self):
-        """
-        Main experiment box.
-        """
-        return self._exp_box
-
-    @property
-    def message_box(self):
-        """
-        Main message box.
-        """
-        return self._message_box
-
-    @property
-    def fitter(self):
-        """
-        Main fitter object.
-        """
-        return self._fitter
+import sys, pkg_resources, pickle
 
 
 class MainWindow(QMainWindow):
@@ -232,12 +89,12 @@ class MainWindow(QMainWindow):
 
         save_fitter = QAction("Save Fitter", self)
         save_fitter.setShortcut("Ctrl+Shift+S")
-        save_fitter.triggered.connect(self.save_fitter)
+        save_fitter.triggered.connect(self.open_save_fitter_dialog)
         file_menu.addAction(save_fitter)
 
         open_fitter = QAction("Open Fitter", self)
         open_fitter.setShortcut("Ctrl+O")
-        open_fitter.triggered.connect(self.open_fitter)
+        open_fitter.triggered.connect(self.open_open_fitter_dialog)
         file_menu.addAction(open_fitter)
 
         file_menu.addSeparator()
@@ -322,18 +179,6 @@ class MainWindow(QMainWindow):
             self._fit_options_dialog.options_signal.connect(self._gui_master.update_fit_options)
             self._fit_options_dialog.show()
 
-    def add_fitter(self):
-        """
-        save fitter to list for use in aic-test
-        """
-        text, ok = QInputDialog.getText(self, 'Save Fitter', 'Enter Name:')
-
-        # save deepcopy of fitter
-        if ok:
-            self._fitter_list[text] = copy.deepcopy(self._fitter)
-            print("Fitter " + text + " saved to list. Current List: ")
-            print(self._fitter_list)
-
     def new_session_callback(self):
         """
         clear everything and start over
@@ -345,7 +190,7 @@ class MainWindow(QMainWindow):
         else:
             pass
 
-    def save_fitter(self):
+    def open_save_fitter_dialog(self):
        """
        save a global_fit object
        """
@@ -355,7 +200,7 @@ class MainWindow(QMainWindow):
        except:
            print("fit not saved")
  
-    def open_fitter(self):
+    def open_open_fitter_dialog(self):
        """
        open a saved global_fit object
        """
