--- pytc_gui/visual/__init__.py
@@ -1,11 +1,0 @@
-__description__ = \
-"""
-visual.py
-
-Submodule holding the major visual elements of the GUI.
-"""
-__author__ = "Michael J. Harms"
-__date__ = "2017-06-01"
-
-from .plot_box import PlotBox
-from .exp_box import ExperimentBox
